fix: renamed detect language and get_llm_response
