+++ src/bot/conversation.py
@@ -2,8 +2,8 @@
 This file contains the functions to be called by the bot that are used to start a conversation with the user and get responses from OpenAI
 """
 from src.repository.vocabulary import get_all_words, get_or_create_user
-from src.services.get_llm_response import llm_response, translate_to_language
-from src.services.detect_language import language_name_from_code
+from services.llm_service import llm_response, translate_to_language
+from services.language_service import language_name_from_code
 
 
 def initializeConversation(message, bot):
